added Puppies component for the Puppies Route, PuppyList is a dumb component that makes singPuppyListItem for each puppy in the array passed to it
