+++ client/components/index.js
@@ -10,3 +10,4 @@ export {Cart} from './cart-table'
 export {default as PuppyList} from './PuppyList'
 export {default as singlePuppyListItem} from './singlePuppyListItem'
 export {ConnectedForm} from './form'
+export {default as Puppies} from './Puppies'
