+++ client/store/index.js
@@ -14,3 +14,4 @@ const store = createStore(reducer, middleware)
 
 export default store
 export * from './user'
+export * from './puppy'
